// src/api/generated/models/postCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItem.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { PostCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItemData } from './postCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItemData';
import type { PostCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItemParsed } from './postCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItemParsed';

export type PostCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItem = {
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  rowNumber: number;
  data: PostCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItemData;
  errors: string[];
  parsed?: PostCompaniesCompanyIdImportsRequestLinesCsvPreview200PreviewRowsItemParsed;
};